fix: company social block
