+++ server/utils/company/company.interface.ts
@@ -1,7 +1,6 @@
-import type { SelectInput, SelectInput } from '../base.interface'
+import type { SelectInput } from '../base.interface'
 import { Company } from './company.model'
 
 export interface ICompanyRepository {
-  selectCompanyCards(config: SelectInput<{}>): Promise<Company[]>
   selectCompanyById(config: SelectInput<{}>): Promise<Company>
 }
