+++ server/utils/company/company.repository.ts
@@ -2,21 +2,10 @@ import { BaseRepository } from '../base.repository'
 import type { SelectInput } from '../base.interface'
 import { ICompanyRepository } from './company.interface'
 import { Company } from './company.model'
-import { CompanyDTO } from './company.dto'
 
 export class CompanyRepository extends BaseRepository<Company> implements ICompanyRepository {
-  dto = new CompanyDTO()
   constructor() {
     super({ loggerPrefix: 'CompanyRepository', Model: Company })
-  }
-
-  async selectCompanyCards(input: SelectInput<{}>): Promise<Company[]> {
-    this.logger.info('selectCompanyCards')
-
-    const data = await this.selectMany<'companies'>(input)
-
-    this.logger.info(`have data ${JSON.stringify(data)}`)
-    return data.map((company) => new Company(company))
   }
 
   async selectCompanyById(config: SelectInput<{}>): Promise<Company> {
@@ -42,6 +31,5 @@ export class CompanyRepository extends BaseRepository<Company> implements ICompa
     })
 
     this.logger.info(`have data ${JSON.stringify(data)}`)
-    return this.dto.validateAndFormatData({ data: new Company(data), dto: config.dto })
   }
 }
